emotion_recognition_app/requirements.txt
Flask
opencv-python
deepface
tensorflow
pillow 
super-gradients
